move tests to fallback_contract
