--- examples/reentrancy/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "test"
-version = "4.0.0-beta"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-publish = false
-
-[dependencies]
-ink = { path = "../../crates/ink", default-features = false }
-main_contract = { path = "main_contract", default-features = false, features = ["ink-as-dependency"] }
-fallback_contract = { path = "fallback_contract", default-features = false, features = ["ink-as-dependency"] }
-
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
-
-[lib]
-name = "test"
-path = "lib.rs"
-crate-type = ["cdylib"]
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-    "main_contract/std",
-    "fallback_contract/std",
-]
-ink-as-dependency = []
-
-[workspace]
-members = [
-    "main_contract",
-    "fallback_contract"
-]
